Remove get_all_repos duplication and move repo lists to __init__.py
