+++ src/reichlab_repo_utils/__init__.py
@@ -1,1 +1,103 @@
 __version__ = "0.0.1"
+
+# source: https://docs.google.com/spreadsheets/d/1UaVsqGQ2uyI42t8HWTQjt0MthQJ-o4Yom0-Q2ahBnJc/edit?gid=1230520805#gid=1230520805
+# (any repo with a WILL_BREAK column = FALSE)
+RULESET_REPO_LIST = [
+    "reichlab-python-template",
+    "duck-hub",
+    # "container-utils",
+    # "covidData",
+    # "distfromq",
+    # "docs.zoltardata",
+    # "ensemble-comparison",
+    # "flu-hosp-models-2021-2022",
+    # "flusion",
+    # "forecast-repository",
+    # "gbq_operational",
+    # "genomicdata",
+    # "hub-infrastructure-experiments",
+    # "idforecastutils",
+    # "jacques",
+    # "jacques-covid",
+    # "llmtime",
+    # "malaria-serology",
+    # "predictability",
+    # "predtimechart",
+    # "qenspy",
+    # "qensr",
+    # "rclp",
+    # "sarimaTD",
+    # "sarix-covid",
+    # "simplets",
+    # "timeseriesutils",
+    # "variant-nowcast-hub",
+    # "Zoltar-Vizualization",
+    # "zoltpy",
+    # "zoltr",
+]
+
+# source: https://docs.google.com/spreadsheets/d/1UaVsqGQ2uyI42t8HWTQjt0MthQJ-o4Yom0-Q2ahBnJc/edit?gid=1230520805#gid=1230520805
+# (any repo with candidate_for_archive column = TRUE)
+ARCHIVE_REPO_LIST = [
+    "duck-hub",
+    # "ensemble-comparison",
+    # "Zoltar-Vizualization",
+    # "container-demo-app",
+    # "2017-2018-cdc-flu-contest",
+    # "2018-2019-cdc-flu-contest",
+    # "activemonitr",
+    # "adaptively-weighted-ensemble",
+    # "ALERT",
+    # "annual-predictions-paper",
+    # "ardfa",
+    # "article-disease-pred-with-kcde",
+    # "bayesian_non_parametric",
+    # "casebot",
+    # "cdcfluforecasts",
+    # "cdcfluutils",
+    # "cdcForecastUtils",
+    # "covid-hosp-forecasts-with-cases",
+    # "covid19-ensemble-methods-manuscript",
+    # "covid19-forecast-evals",
+    # "d3-foresight",
+    # "dengue-data-stub",
+    # "dengue-ssr-prediction",
+    # "dengue-thailand-2014-forecasts",
+    # "densitystackr",
+    # "diffport",
+    # "ensemble-size",
+    # "flu-eda",
+    # "flusight-csv-tools",
+    # "Flusight-forecast-data",
+    # "FluSight-package",
+    # "flusight-test",
+    # "flusurv-forecasts-2020-2021",
+    # "forecast-framework-demos",
+    # "forecastTools",
+    # "foresight-visualization-template",
+    # "german-flu-forecasting",
+    # "hubEnsembles",
+    # "kcde",
+    # "ledge",
+    # "lssm",
+    # "make-example",
+    # "mmwr-week",
+    # "mvtnorm-mod-kcde",
+    # "ncov",
+    # "neural-stack",
+    # "nuxt-forecast-viz",
+    # "online-lag-ensemble",
+    # "pdtmvn",
+    # "pkr",
+    # "proper-scores-comparison",
+    # "pylssm",
+    # "reviewMID",
+    # "shiny-predictions",
+    # "ssr-influenza-competition",
+    # "style",
+    # "tracking-ensemble",
+    # "TSIRsim",
+    # "xgboost-mod",
+    # "xgbstack",
+    # "xpull",
+]
